VC warning fixes
git-svn-id: svn://svn.code.sf.net/p/actiongame/code/trunk@1747 e016b1c5-ed68-44aa-a2f7-31cac88cffde

+++ ac/source/src/server.cpp
@@ -280,7 +280,7 @@ void sendpacket(int n, int chan, ENetPacket *packet, int exclude = -1)
 {
     if(n<0)
     {
-        recordpacket(chan, packet->data, packet->dataLength);
+        recordpacket(chan, packet->data, (int)packet->dataLength);
         loopv(clients) if(i!=exclude) sendpacket(i, chan, packet);
         return;
     }
